+++ CHANGELOG.md
@@ -45,6 +45,8 @@ IMPROVEMENTS:
   - Detect PuTTY Link SSH client on Windows and show an error. [GH-1518]
   - `vagrant ssh` in Cygwin won't output DOS path file warnings.
   - Add `--rtcuseutc on` as a sane default for VirtualBox. [GH-912]
+  - SSH will send keep-alive packets every 5 seconds by default to
+    keep connections alive. Can be disabled with `config.ssh.keep_alive`. [GH-516]
 
 BUG FIXES: